// src/apis-client/models/ExchangeLocationInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aeler demo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ExchangeLocationInput
 */
export interface ExchangeLocationInput {
    /**
     * 
     * @type {string}
     * @memberof ExchangeLocationInput
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ExchangeLocationInput
     */
    depotId?: string;
    /**
     * 
     * @type {string}
     * @memberof ExchangeLocationInput
     */
    addressId?: string;
    /**
     * 
     * @type {string}
     * @memberof ExchangeLocationInput
     */
    locationId?: string;
}

export function ExchangeLocationInputFromJSON(json: any): ExchangeLocationInput {
    return ExchangeLocationInputFromJSONTyped(json, false);
}

export function ExchangeLocationInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExchangeLocationInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'depotId': !exists(json, 'depotId') ? undefined : json['depotId'],
        'addressId': !exists(json, 'addressId') ? undefined : json['addressId'],
        'locationId': !exists(json, 'locationId') ? undefined : json['locationId'],
    };
}

export function ExchangeLocationInputToJSON(value?: ExchangeLocationInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'depotId': value.depotId,
        'addressId': value.addressId,
        'locationId': value.locationId,
    };
}